+++ CHANGELOG.md
@@ -7,12 +7,15 @@
 - feat: Add `pydalec-test` TCP CLI
 - feat: Add method to fetch current GNSS location
 - feat: add acquiring location to CLI test script
+- feat: Add method to fetch current sun zenith
 
 ### Changed
 
 - fix: `Transport.connected` is single source of truth for connection state
 - refactor: rewrite measurement buffer handling
-- feat: Track measurements by object reference for efficient polling
+- refactor: Track measurements by object reference for efficient polling
+- feat: CLI also reports current sun zenith
+- feat: Add multiline `Measurement.__str__` format
 
 ### Fixed